chore: Unify the `MERKLE_TREE_HEIGHT` constant (#292)
Use it both for Transaction and Event Merkle Tree.

+++ system-programs/programs/merkle_tree_program/src/transaction_merkle_tree/update_merkle_tree_lib/merkle_tree_update_state.rs
@@ -20,7 +20,7 @@ pub struct MerkleTreeUpdateState {
     pub current_level: u64,
     pub current_level_hash: [u8; 32],
     pub tmp_leaves_index: u64,
-    pub filled_subtrees: [[u8; 32]; MERKLE_TREE_HEIGHT as usize],
+    pub filled_subtrees: [[u8; 32]; MERKLE_TREE_HEIGHT],
 
     pub leaves: [[[u8; 32]; 2]; 16],
     pub number_of_leaves: u8,
